Add files via upload
This is a bash script which uses some of the major security tools used for bug bounty. And to automate everything.
Tools Used:- 
"Subfinder", "Assetfinder" for subdomain enumeration.
"Httpx" for live domains.
"Naabu" for port scanning.
"Nuclei" for cve , exploits scan.
"Param Spider" to find parameters for xss and sql injection bugs.
"sqlmap" for sql injection.

This tool will save the output by creating a directory by the name of the website we are scanning.